업데이트/회원가입및로그인 : 회원가입 페이지와 로그인 페이지 개발
Just-Validate.js와 bulma.css를 이용.
Just-Validate로 회원가입과 로그인 기능 시 form검증.
bulma.css로 html스타일 추가
시큐리티 콘피그에서 로그인 검증시 커스텀 예외 처리.
로그인 성공과 실패 핸들러 클래스 작성
멤버엔티티 개인정보수집이용 필드 추가

+++ src/main/java/com/scit/iLog/domain/sentimentalAnalysis/WeatherEntity.java
@@ -23,13 +23,13 @@ public class WeatherEntity extends BaseTimeEntity {
     private AnalysisTargetEntity analysisTarget;
 
     @Column(name = "temperature")
-    private float temperature;
+    private double temperature;
 
     @Column(name = "humidity")
     private int humidity;
 
     @Column(name = "wind_speed")
-    private float windSpeed;
+    private double windSpeed;
 
     @Column(name = "weather_desc")
     private String weatherDesc;
